WIP autocomplete for JSX props.
Autocomplete for JSX props is only distinguished from autocomplete for identifiers by the context in which it takes place. Compared to completion of labeled arguments, it is not triggered by any special characters such as `~`.

The context is used as a heuristic to interpret the **intention**.

So for example, this should complete as a prop from `Comp`:

```rescript
<Comp foo
```

but this should complete as an identifier:

```rescript
<Comp x=foo
```

this should also complete as an identifier:

```rescript
<Comp List.m
```
as even though it's in a place where a prop name is expected, the intent is clearly to complete a value from module `List`.

This is a WIP to define what the contexts are.